requirements.txt
akshare>=1.14.0
pandas>=2.0.0
ta>=0.11.0
google-generativeai>=0.7.0
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0